requirements-test.txt
pytest>=8.0
pytest-django>=4.8
pytest-cov>=5.0        # relatório de cobertura (recomendo)